# sisged-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SISGED API",
        "description": "API REST do sistema SISGED. Documentação gerada com L5-Swagger (OpenAPI 3).",
        "contact": {
            "email": "contato@sisged.com.br"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Servidor da API"
        }
    ],
    "paths": {
...
        "/alunos": {
            "get": {
                "tags": [
                    "Alunos"
                ],
                "summary": "Lista todos os alunos",
                "description": "Retorna uma lista paginada de alunos cadastrados",
                "operationId": "df60868136a72dcde3f23d5761e263a7",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Quantidade de registros por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de alunos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Aluno"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Alunos"
                ],
                "summary": "Cadastra um novo aluno",
                "operationId": "40440e954bbf3923a93be1388ac34755",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nomeAluno",
                                    "cpfAluno",
                                    "emailAluno"
                                ],
                                "properties": {
                                    "nomeAluno": {
                                        "type": "string",
                                        "example": "Maria Silva"
                                    },
                                    "cpfAluno": {
                                        "type": "integer",
                                        "example": 12345678900
                                    },
                                    "emailAluno": {
                                        "type": "string",
                                        "example": "maria.silva@email.com"
                                    },
                                    "telefoneAluno": {
                                        "type": "integer",
                                        "example": 31998765432
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Aluno criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Aluno"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/alunos/{idAluno}": {
            "get": {
                "tags": [
                    "Alunos"
                ],
                "summary": "Exibe os dados de um aluno específico",
                "operationId": "8639853b5de3ead7eb876dd4d2b398e5",
                "parameters": [
                    {
                        "name": "idAluno",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aluno encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Aluno"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aluno não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Alunos"
                ],
                "summary": "Atualiza os dados de um aluno",
                "operationId": "3a909c17b258c9b4673046cf6a59c176",
                "parameters": [
                    {
                        "name": "idAluno",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nomeAluno": {
                                        "type": "string",
                                        "example": "Maria Silva Souza"
                                    },
                                    "cpfAluno": {
                                        "type": "integer",
                                        "example": 12345678900
                                    },
                                    "emailAluno": {
                                        "type": "string",
                                        "example": "maria.souza@email.com"
                                    },
                                    "telefoneAluno": {
                                        "type": "integer",
                                        "example": 31998765432
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Aluno atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Aluno"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aluno não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Alunos"
                ],
                "summary": "Remove um aluno",
                "operationId": "29e37d299e3f6b26c64b40c68dbdb866",
                "parameters": [
                    {
                        "name": "idAluno",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Aluno removido com sucesso"
                    },
                    "404": {
                        "description": "Aluno não encontrado"
                    }
                }
            }
        },
...
        "/auth/administrador/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Login de administrador (acesso total)",
                "operationId": "5d03c7a8e92386f10a9ea4afc0d92dbf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "emailAdministrador",
                                    "senhaAdministrador"
                                ],
                                "properties": {
                                    "emailAdministrador": {
                                        "type": "string",
                                        "example": "admin@sisged.com"
                                    },
                                    "senhaAdministrador": {
                                        "type": "string",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso"
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/auth/aluno/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Login de aluno (acesso restrito, somente leitura de cursos)",
                "operationId": "d41ddf7aa75e9d868e7e8b4e69c9b84c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "emailAluno",
                                    "senhaAluno"
                                ],
                                "properties": {
                                    "emailAluno": {
                                        "type": "string",
                                        "example": "aluno@email.com"
                                    },
                                    "senhaAluno": {
                                        "type": "string",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso"
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Logout (invalida o token atual)",
                "operationId": "6abf3b64a4bc7838d56346f05a5153af",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Aluno": {
                "title": "Aluno",
                "required": [
                    "nomeAluno",
                    "cpfAluno",
                    "emailAluno"
                ],
                "properties": {
                    "idAluno": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "nomeAluno": {
                        "type": "string",
                        "example": "Maria Silva"
                    },
                    "cpfAluno": {
                        "type": "integer",
                        "format": "int64",
                        "example": 12345678900
                    },
                    "emailAluno": {
                        "type": "string",
                        "format": "email",
                        "example": "maria.silva@email.com"
                    },
                    "telefoneAluno": {
                        "type": "integer",
                        "format": "int64",
                        "example": 31998765432,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}